Make InterningBinaryReader pool buffers (#6556)
Fixes #3954

Context
The Buffer class is 100 kB (40 for characters and 60 for the byte array) and it is being allocated a lot, easily in the order of tens of thousands of instances per build. OOM crashes are often blamed on such allocations.

Changes Made
Made the byte array allocation lazy since in many scenarios it ends up not being used.
Implemented a trivial Buffer pooling which eliminates >97% of allocations.