Fix missing Reference issue
Some references can be resolved later. I'd initially missed it because References wasn't invoked directly, but rather a Reference was retrieved then later modified. I believe this is the only place that happens, however. The first check essentially checks whether it has already been resolved, and the second place where the ChangeWave was enabled ensures it is canonicalized.