Simplify WorkloadResolver.GetInstalledManifests
We now have a collection of the manifests that are loaded, so use
that rather than rescanning. This makes it behave correctly for
correct for overlay resolvers.

Also return an "Info" object to be consistent with other API. We
may want to add more information onto this object in future.