VS 1449000: Fix handling of satellite assemblies in ClickOnce (#6665)
Fixes AB#1449000

### Context
Satellite assemblies are not being handled correctly after ClickOnce changes for .NET Core publishing were made in VS 16.8.

Following issues have been observed post those changes:

1. If packages publish en-us resource assembly for other locales as well, we will end up adding all of them in the ClickOnce manifest. Since the strong name signature will be identical for all these resource assemblies, we will now have multiple assemblies with identical signature in the ClickOnce manifest. This will cause CO Install to fail.

2. The References item list can also contain resource assemblies. If these assemblies are also present in the SatelliteAssembly item list, we will now end up writing 2 entries for the same assembly in the manifest. This will cause CO install to fail.

### Changes Made
When we process the References item list, we will maintain a dictionary for satellite assemblies that are getting included in the list of assemblies. Later when we process the SatelliteAssemblies item list, we will look up in this dictionary and skip the entries that are already included through the References list.