Fix lock contention in ProjectRootElementCache.Get (#6680)
Fixes #3039

### Context
There is significant amount of lock contentions in ProjectRootElementCache.Get method. 
(For OrchardCore project, for example, the total time spent waiting to take a lock ~1sec.)
Reason: the code is using a single lock to control to access the cache, and also read file into XmlDocument, when it does not exist. The later one can be slow on a slow disk.

### Changes Made
The ProjectRootElementCache.Get function is rewritten to exclude IO operations from the locked sections.
For this goal we also had to separate logic of loading from logic of adding the element to the cache.

### Testing
Unit tests + manual testing.