Miscellaneous event tweaks (#6725)
* Make EvaluateStop event adjacent to EvaluateStart

It's not well documented, but EventSource requires that the start and stop event IDs
for a single thing be separated by exactly one. Since this one didn't, we weren't
getting DURATION_MSEC computed for overall evaluation time in our traces.

* Update Evaluate* event doc comments

* Include EvaluateStart/Stop in DOTNET_PERFLOG

This is a more critical measure than Parse, which is XML->memory, because it
includes imports, SDK resolution, and MSBuild logic.

* Drop Parse from DOTNET_PERFLOG

It's pretty rare that the XML parse phase of project load is a signifcant
part of evaluation time; it's usually a tiny fraction of overall evaluation.