Update schema for combining TargetFramework info to allow for invalid xml names such as including '+' (#6699)
Fixes #6603

Context
The ability to gather extra properties from referenced projects was added in #5994. It used XML transformations to couch that logic in a way that MSBuild could understand.

Unfortunately, the XML-based logic assumed everything was a valid XML element. Ultimately, the goal of the change, at least as I understand it, was just to pass information.

Changes Made
TargetFrameworks had been XML elements, but some are not valid XML names. This was changed to being a property on the more generically named TargetFramework element. This also allows for escaping.

Testing
Verified that a simple case that had not previously worked now works and looked at what it produced in a debugger.

Notes
After this PR, the SDK can opt in. Then MSBuild can turn it on by default, and the SDK can stop specifying the property.