Help GenerateResource find assemblies
If GenerateResource is forced to create a new App Domain, it loses modifications made to the parent app domain. With mainline MSBuild, this is no problem because we don't make modifications. With MSBuildLocator, we do, and it prevents it from finding the assemblies it needs to execute this task.

This adds an extra "Setup" method before we use the new app domain that adds assemblies next to MSBuild to the app domain.