Increase ProjectRootElementCache MRU cache
The average imports/project in a recent OrchardCore solution
(primarily ASP.NET Core) is 83, so one project can by itself
evict stuff from the old cache of size 50.

There are likely further improvements to be made in this area
(some ideas in #6551), but increasing the size to 200 seems to
be a significant improvement with an extremely small change.