Add InternalsVisibleTo to common types schema (#6778)
Fixes #6777

Since .NET 5, the <InternalsVisibleTo Include="MyProject.Assmebly" /> item type has been supported to generate the System.Runtime.CompilerServices.InternalsVisibleTo for the output assembly. The definition for this item type should be in the common types schema file so that editors (like Visual Studio) provide statement completion and a QuickInfo tooltip for it.