Switch to full NGEN (#6764)
Context
We are currently pre-compiling only code that gets executed as part of our IBC training scenarios. This results in smaller native images but the coverage is not perfect and it's easy to miss a code path and cause JITting at run-time.

Changes Made
With Visual Studio switching to 64-bit, address space is no longer a concern and the positive impact of pre-compiling everything outweighs the cost of larger image sizes.

Testing
Experimental VS insertion showing improvements in # of methods JITted and wall-clock time.