Only set debug path when MSBuildDebugEngine is set (#6792)
In #6639 I thought it would be a good idea to always expose exceptions by always setting msbuild's debug path, fallbacking to the current working directory. But that's a breaking change.

Also, some processes (like VS helpers) do not propagate the user's CWD and thus end up writing under Program Files which fails with permission exceptions.

So only set the debug path to the current working directory when `MSBuildDebugEngine` is set.