Expose LogTaskInputs to tasks (#6803)
Fixes #6305

Context
The LogTaskInputs flag passed to MSBuild in BuildParameters can be used by tasks to optimize logging. When it's true, a task can omit its own duplicated logging of inputs.

Changes Made
LogTaskInputs is plumbed through and exposed as IsTaskInputLoggingEnabled on EngineServices and also on TaskLoggingHelper for convenience. The RAR task is updated to use the new property.

Testing
New unit test.
Manual testing with and without /v:diag and inspecting the log.