Optimize InternableString.GetHashCode (#6816)
Context
The current straightforward implementation is slow, especially on 64-bit Framework CLR. It is causing a measurable regression when evaluating large projects.

Changes Made
Rewrote the routine using a similar approach as this BCL method. Calculating only 2 and not 4 characters at a time, though, to reduce the complexity since our implementation works on a list of spans and not just one string. The additional perf benefit of going 4 at a time would be relatively small.

Testing
A new unit test to verify correctness.
Micro-benchmark showing 1.6x boost on x86 and 2x boost on x64.
Evaluation perf traces showing significant reduction in the time spent in GetHashCode, saving 110 ms (~6%) per evaluation of the Unreal Engine C++ project on 64-bit.