Update BlazorWebAssembly targets to work with updated wasm targets
[This PR](https://github.com/dotnet/runtime/pull/57556) adds support for
relinking the native wasm files during the build (instead of only during
publish).

This required changes to the targets available, and how they can be
used, and thus require changes in the blazor targets to work with that.

The new setup is explained in
https://github.com/radical/runtime/blob/wasm-build/src/mono/wasm/build/README.md .

Forward ports the changes from release/6.0.1xx