Log message arguments for warnings and errors (#6804)
We were only logging the arguments for messages, but not for errors or warnings. Fortunately we can fix this and avoid incrementing the binlog format version.

First this moves the logic to set the Arguments flag into a common codepath close to Message, and only sets the flag when the arguments are present and we're either a message, warning or error. Previously it was only set for messages (and all types derived from Message).

Technically all BuildStatusEventArgs and CustomBuildEventArgs also inherit from LazyFormattedBuildEventArgs, but the serialization for them is fully manual and we don't even write the message for them at all. We ensure that if by any accident an instance of BuildStatus or some other random type such as CustomBuildEventArgs sets the Arguments, we don't set the flag because we're not going to be writing arguments for those.

We extract a method to log the arguments, and call it for errors and warnings.

Note that this doesn't change the BuildEventArgsReader at all, which proves that a file format version change is not necessary.

This also ensures that the tests for error, warning and message hit both codepaths, with and without arguments.

Fixes https://github.com/dotnet/msbuild/issues/6790.