Add enumeration values for DebugType in XSD (#6849)
Context
The `DebugType` property has a set of expected values, and is not just a plain text property.
Currently completion on the `DebugType` property offers no guidance.

Changes Made
This change adds those values so they appear in completion. It also documents that `pdbonly` is equivalent to `full` in recent compilers.