Pass and undefine Platform property
Instead of a mix between PlatformTarget and Platforms.
This fixes an issue with legacy projects that define values
for OutputPath and PlatformTarget based on the Platform|Configuration.
It also fixes an issue where a user can start a command line build by
passing /p:Platform=foo, and that global property would carry to ALL
p2p hops. This can lead to weird issues such as the Platform being x86
when a p2p hop might pass PlatformTarget=AnyCPU