Remove O# workaround from RazorSDK (#21806)
As part of addressing https://github.com/dotnet/aspnetcore/issues/30750, we had added
a workaround to the RazorSDK to ensure building inside O# worked. This issue is being
resolved as part of the upcoming O# v1.37.16. In the meanwhile, the workaround affects
sourcelink for razor files.

Fixes https://github.com/dotnet/aspnetcore/issues/37182