Update tests to not assume there's a 4.0 runtime on the computer.
VS 2022 doesn't have it, so updating to that pool requires not assuming
it. This should be more stable but it now has a variable as its
"TargetFrameworkVersion"
(MSBuildConstants.StandardTestTargetFrameworkVersion) in case we need to
change it. Also cleaned up some Tools="" in tests and made some other
simplifications. The only part that should have any product effect is in
ToolLocationHelper where, if we had previously allocated a new empty List to
replace an empty List, we just used that other list.