Allow CopyToOutputDirectory metadata as attribute
The XSD currently expects known item metadata to be specified using nested XML elements. However metadata may also be specified as XML attributes on the item's element. This is a common pattern, especially when hand editing.

Here we add schema items for the CopyToOutputDirectory metadata. We could extend this to other metadata values.