Add enumeration values for CopyToOutputDirectory in XSD (#7024)
* Add enumeration values for CopyToOutputDirectory in XSD

This property only supports values Never, Always or PreserveNewest. This change allows VS and other consumers of the XSD to suggest and validate that the provided value is valid.

The property on some item types was previously declared incorrectly as being a boolean.

* Allow CopyToOutputDirectory metadata as attribute

The XSD currently expects known item metadata to be specified using nested XML elements. However metadata may also be specified as XML attributes on the item's element. This is a common pattern, especially when hand editing.

Here we add schema items for the CopyToOutputDirectory metadata. We could extend this to other metadata values.