Ship System.ValueTuple.dll to VS
This file is not used in normal MSBuild operation because Visual Studio
and MSBuild both have binding redirects to use the version from the GAC
in .NET Framework, but if an API client application doesn't have such a
binding redirect, they might fail. This keeps users of MSBuildLocator
working by shipping the file for their use.

Fixes #6976.