Remove type casts
The use of `as` suggests some kind of runtime type checking is involved, and that it may fail. In fact, the type is statically known to implement the interface, so direct assignment is sufficient.

There should be no runtime change here as the compiler appears to optimise the check away. I just find this form of the code clearer.