Introduce Dependabot (#7046)
Guidance from dnceng is to use Dependabot instead of Darc/Maestro to manage updating packages to non-.NET or GA versions of packages (anything that's not "the current development build").

This commit introuduces Dependabot and moves some of our dependencies to be managed by it--specifically the ones that are build/test-time only and the ones that are embedded into our outputs. We should have free reign to control the versions of those any time without introducing downstream dependency problems.