Remove locking from CopyOnWritePropertyDictionary<>
This commit removes all `Monitor`-based locking from `CopyOnWritePropertyDictionary<>`.

To achieve this, the backing store has been changed from `CopyOnWriteDictionary<T>` to `ImmutableDictionary<string, T>`. Note that the former was just a thin wrapper around the latter, so this change inlines a layer of indirection.

With the immutable collection available in scope, the members of this type can be implemented so that locking is not needed.