Change messages for ILLink (#22636)
Changes messages for ILLink:
We now print "Optimizing assemblies for size may change the behavior of the app. Be sure to test after publishing. See: https://aka.ms/dotnet-illink" in the case the warnings are suppressed. To let people know that trimming with suppressed warnings can be dangerous. Also, this message is printed before running trimming, instead of after trimming

We now print "Optimizing assemblies for size. This process might take a while." before running trimming. In incremental builds that trimming is not run, we don't print this message.

Add test for messages

Add net7 to test utils