Al.exe finds the correct tool based on platform (#7051)
Finally puts https://github.com/dotnet/msbuild/issues/5981 to bed.
(to trigger an auto-close: fixes https://github.com/dotnet/msbuild/issues/5981)

### Context
The underlying problem with the Al issues is that **the Al task finds its own tool based on the current process' architecture**. This means:

x86 msbuild.exe -> x86 Al.exe
x64 msbuild.exe -> x64 Al.exe

Now, if your x86 msbuild.exe is building with `Platform=x64`, the x86 Al.exe gets passed `platform:x64` and logs its own warning because of mismatched platforms.

So we fixed that by checking if the platform was x64, then look in the x64 tools directory in common.targets.

Now we're hitting problems where the x64 msbuild.exe is calling the x64 Al.exe and being passed `platform:x86`, causing the reverse of the original issue!

### Changes Made
The Al task checks the platform that was passed. If it's x86, it will find the 32 bit al.exe. If x64, it will append x64 to the path before finding the tool.

This also reverts appending x64 to the tools directory in common.currentversion.targets before calling the Al task. It shouldn't have worked in today's x64 msbuild.exe, but does because of the x86 fallback behavior. Apending x64 before the task is called is no longer required.

### Testing

### Notes
I allowed `_ALExeToolPath` to be overridden to account for projects that may be using it today with older msbuild binaries.

**A mismatched common.currentversion.targets & microsoft.build.dll can fail**.

Old targets & new dll: if platform is x64 it will double append x64, but the fallback will find the x64 al.exe. no fix required. if platform is x86 and x64 is appended to the path: will likely find the x64 tool and log the error. fix: customer should no longer append x64 to the path.

new targets & old dll: for platform=x64, `x64` needs to be manually appended. The "source of truth" is setting `AlToolPath` to end in `x64`. for platform=x86, no fix needed.