Replace O(N^2) with O(N) (#7054)
This loop verifies all items in one list are present in another. Previously this involved ArrayList.Contains, which is O(N). As this happened in a loop, the time complexity comes out to O(N^2). Using a HashSet<> instead allows O(1) lookup within the loop, leading to an overall complexity of O(N).