Minor refactor around project cache logging (#7098)
The overall goal I'm trying to achieve is general logging cleanup of the project cache code. Currently it looks nothing like regular MSBuild logging, isn't associated with projects, and various other quirks.

This specific change is a light refactor of how the `PluginLoggerBase` provided to plugins is created. It allows for eventually providing the `BuildEventContext`.

It also removes an unused Verbosity provided to the logger. This is technically a breaking change to the public API, although a) `PluginLoggerBase` is in the experimental namespace, b) project cache is new and not likely to be widely used, and c) callers have no way currently to provide their own instance of `PluginLoggerBase`, so subclassing it in the first place makes no sense and thus is very likely not done outside of maybe unit tests.