Enable nullable reference types by default
This turns C# 9's nullable reference types on by default for all projects in the solution, and opts out unannotated files by adding `#nullable disable` to them.

This has two nice properties:

1. New code will be written with nullability in mind as the feature will be on by default.
2. It's easy to find unannotated code by searching for `#nullable disable`.

We have taken this approach in the .NET Project System and in CPS, and it's worked very well for us.

Code under `Deprecated` is unchanged.