Avoid loading Framework/Utilities in CodeTaskFactory (#7129)
Fixes #7108 by moving the check of whether a given assembly is one of
the ones we'll provide by default to occur before the assembly is
loaded. This should prevent loading the amd64 copy of either assembly
side-by-side with the bin copy, and thus keep downstream consumers
from getting type definitions from the wrong assembly.