Enable solution filtering with -graph option (#7084)
Fixes #7058

Context
Solution filters operate in MSBuild by modifying the list of projects that "should build." This list is normally constructed in SolutionProjectGenerator, but for -graph builds, it's in GraphBuilder instead. This was hard to notice because it didn't actually cause a build failure because the slnf tries to get out of the way as quickly as it can when you try to build it, which meant the graph build simply built the full sln. This corrects the slnf filtering behavior with -graph.

Changes Made
Filter the list of projects that "should build" when using -graph.

Testing
I built MSBuild.dev.slnf before and after the change as well as with and without graph. Eyeballing it, there were about the same number of projects built with the change and -graph as without the change and without -graph. There were substantially more with -graph but without the change.