Fix bug in CacheSerialization.SerializeCaches (#7140)
Context
The method CacheSerialization.SerializeCaches was using File.OpenWrite(fullPath) which is pretty confusing if a file already exists, as it will write to it without truncating it. So if a cache file was 10MB and we then write a file that is just 5KB, the file will still be 10MB.

Changes Made
Use FileStream constructor instead with a more adequate behavior.