Fix stackoverflowexception (#7053)
Fixes #6925

Context
If you have a cycle in your ProjectReferences, that should be caught and marked as invalid with MSB4251. If you do a graph build, however, it turns into a StackOverflowException, which is much less clear. This handles circular dependencies, giving the nicer exception.

Changes Made
Ensures that each project "marks" itself as not to be revisited before the recursive call so that you cannot get endless recursion.

Testing
Tried with the original repro and got MSB4251 as desired.

Notes
The first commit maintains the previous caching behavior. It isn't obvious to me how much benefit there really is here, and I have yet to test it. In particular, although it could save you a good bit of work, it costs creation and manipulation of a series of intermediate HashSets that are all stored in a Dictionary, so GC can't even get rid of any of them. Without having tested this, I would guess that the first commit would be more efficient for large projects, and the second commit would be more efficient for small projects. I'm happy to hear feedback before I test or just test that hypothesis as reviewers think best.