Port doc comment changes from docs repo (#7152)
Fixes #7151

Context
Doc comments are changed from time to time in the dotnet-api-docs repo. These changes need to be ported back to the MSBuild source, now that the MSBuild source is being used as the source-of-truth for MSBuild reference docs.

Changes Made
Comment-only changes in a number of source files.

Testing
Notes
The changes that were required can also be found here: dotnet/dotnet-api-docs#7484 - this docs PR was necessary to prevent the dotnet-api-docs changes from being overwritten when the comments were re-read from MSBuild sources for the 17.0 version of MSBuild.