Validate binding redirects (#7153)
Fixes #7078

Context
If our binding redirects are off from the packages we actually ship, that can cause problems downstream when we try to integrate with anything other than MSBuild. This ensures they are correct with certain exceptions specified in a comment.

Changes Made
Added a post-build (and tests) validation step.