Remove CultureName and CultureInfo from LazyFormattedBuildEventArgs (#7254)
Fixes #7015

Context
The intent was to stash the current culture when the event args are created and then use it for formatting (which may happen later). It turns out that despite the code in WriteToStream and CreateFromStream, the culture doesn't propagate when our args are serialized and deserialized; see how new instances are created during deserialization:

msbuild/src/Shared/LogMessagePacketBase.cs

Lines 517 to 535 in 14efa06

                 LoggingEventType.BuildErrorEvent => new BuildErrorEventArgs(null, null, null, -1, -1, -1, -1, null, null, null), 
                 LoggingEventType.BuildFinishedEvent => new BuildFinishedEventArgs(null, null, false), 
                 LoggingEventType.BuildMessageEvent => new BuildMessageEventArgs(null, null, null, MessageImportance.Normal), 
                 LoggingEventType.BuildStartedEvent => new BuildStartedEventArgs(null, null), 
                 LoggingEventType.BuildWarningEvent => new BuildWarningEventArgs(null, null, null, -1, -1, -1, -1, null, null, null), 
                 LoggingEventType.ProjectFinishedEvent => new ProjectFinishedEventArgs(null, null, null, false), 
                 LoggingEventType.ProjectStartedEvent => new ProjectStartedEventArgs(null, null, null, null, null, null), 
                 LoggingEventType.TargetStartedEvent => new TargetStartedEventArgs(null, null, null, null, null), 
                 LoggingEventType.TargetFinishedEvent => new TargetFinishedEventArgs(null, null, null, null, null, false), 
                 LoggingEventType.TaskStartedEvent => new TaskStartedEventArgs(null, null, null, null, null), 
                 LoggingEventType.TaskFinishedEvent => new TaskFinishedEventArgs(null, null, null, null, null, false), 
                 LoggingEventType.TaskCommandLineEvent => new TaskCommandLineEventArgs(null, null, MessageImportance.Normal), 
 #if !TASKHOST // MSBuildTaskHost is targeting Microsoft.Build.Framework.dll 3.5 
                 LoggingEventType.TaskParameterEvent => new TaskParameterEventArgs(0, null, null, true, default), 
                 LoggingEventType.ProjectEvaluationStartedEvent => new ProjectEvaluationStartedEventArgs(), 
                 LoggingEventType.ProjectEvaluationFinishedEvent => new ProjectEvaluationFinishedEventArgs(), 
                 LoggingEventType.ProjectImportedEvent => new ProjectImportedEventArgs(), 
                 LoggingEventType.TargetSkipped => new TargetSkippedEventArgs(), 
                 LoggingEventType.Telemetry => new TelemetryEventArgs(), 
This, together with @KirillOsenkov's comment in the linked issue, and the general doubt about the incidence of scenarios where this would be useful (a custom task emitting custom build events, explicitly changing the current culture, and relying in the culture-specific formatting) makes me want to take the risk and just remove all of this.

Changes Made
Removed the fields and the related logic. Also removed a couple of redundant CultureInfo.CurrentCulture arguments.

Testing
Built large solutions, verified that the code being removed was not hit/used.

Notes
This PR is changing private members of a serializable (but not ISerializable) type so technically a breaking change. We've done this before when we tweaked fields in the same class in #7010 and serialization is intended for IPC here anyway, but still, if we want to be 100% backward compatible, this PR would have to be rejected.

The perf win in terms of memory allocations is in the order of 0.1's of % in a typical build.